+++ src/main/java/com/sunmi/sunbay/nexus/model/common/PostAuthAmount.java
@@ -21,22 +21,22 @@ public class PostAuthAmount {
     /**
      * Order amount (required)
      */
-    private Double orderAmount;
+    private Integer orderAmount;
 
     /**
      * Tip amount (optional)
      */
-    private Double tipAmount;
+    private Integer tipAmount;
 
     /**
      * Tax amount (optional)
      */
-    private Double taxAmount;
+    private Integer taxAmount;
 
     /**
      * Surcharge amount (optional)
      */
-    private Double surchargeAmount;
+    private Integer surchargeAmount;
 
     /**
      * Pricing currency (ISO 4217, required)
